modified add_course_offering data

+++ schema.sql
@@ -168,7 +168,7 @@ CREATE TABLE Credit_cards (
 /* Package might not be offered but customer should be able to finish their remaining redemptions*/
 CREATE TABLE Buys (
     package_id INTEGER REFERENCES Course_packages ON DELETE SET NULL, 
-    number INTEGER REFERENCES Credit_cards ON DELETE CASCADE,
+    number BIGINT REFERENCES Credit_cards ON DELETE CASCADE,
     b_date TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     num_remaining_redemptions INTEGER NOT NULL,
     PRIMARY KEY (package_id, number, b_date),
